+++ inference-engine/src/legacy_api/include/legacy/ngraph_ops/convolution_ie.hpp
