+++ inference-engine/src/legacy_api/src/transformations/conv_bias_fusion.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "transformations/common_optimizations/conv_bias_fusion.hpp"
+#include <legacy/transformations/conv_bias_fusion.hpp>
 
 #include <memory>
 #include <numeric>
@@ -13,11 +13,11 @@
 #include <ngraph/rt_info.hpp>
 #include <ngraph/pattern/op/wrap_type.hpp>
 
-#include <ngraph_ops/convolution_ie.hpp>
-#include <ngraph_ops/deconvolution_ie.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
+#include <legacy/ngraph_ops/deconvolution_ie.hpp>
 
 #include <transformations/utils/utils.hpp>
-#include "itt.hpp"
+//#include "../../../transformations/src/itt.hpp"
 
 using namespace ngraph;
 
@@ -162,7 +162,7 @@ bool conv_callback(ngraph::pattern::Matcher &m) {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvAddFusion, "ConvAddFusion", 0);
 
 ngraph::pass::ConvAddFusion::ConvAddFusion() {
-    MATCHER_SCOPE(ConvAddFusion);
+    //MATCHER_SCOPE(ConvAddFusion);
     auto conv = ngraph::pattern::wrap_type<op::ConvolutionIE>(pattern::consumers_count(1));
     auto add = ngraph::pattern::wrap_type<opset1::Add>({conv, pattern::any_input()});
 
@@ -177,7 +177,7 @@ ngraph::pass::ConvAddFusion::ConvAddFusion() {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvMultiplyFusion, "ConvMultiplyFusion", 0);
 
 ngraph::pass::ConvMultiplyFusion::ConvMultiplyFusion() {
-    MATCHER_SCOPE(ConvMultiplyFusion);
+    //MATCHER_SCOPE(ConvMultiplyFusion);
     auto conv = ngraph::pattern::wrap_type<op::ConvolutionIE>(pattern::consumers_count(1));
     auto add = ngraph::pattern::wrap_type<opset1::Multiply>({conv, pattern::any_input()});
 
@@ -192,7 +192,7 @@ ngraph::pass::ConvMultiplyFusion::ConvMultiplyFusion() {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::DeconvAddFusion, "DeconvAddFusion", 0);
 
 ngraph::pass::DeconvAddFusion::DeconvAddFusion() {
-    MATCHER_SCOPE(DeconvAddFusion);
+    //MATCHER_SCOPE(DeconvAddFusion);
     auto conv = ngraph::pattern::wrap_type<op::DeconvolutionIE>(pattern::consumers_count(1));
     auto add = ngraph::pattern::wrap_type<opset1::Add>({conv, pattern::any_input()});
