+++ inference-engine/src/legacy_api/src/transformations/convert_convolutions.cpp
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "itt.hpp"
-#include "transformations/op_conversions/convert_convolutions.hpp"
+//#include "../../../transformations/src/itt.hpp"
+#include <legacy/transformations/convert_convolutions.hpp>
 
 #include <memory>
 #include <vector>
@@ -11,8 +11,8 @@
 #include <ngraph/opsets/opset1.hpp>
 #include <ngraph/rt_info.hpp>
 
-#include <ngraph_ops/convolution_ie.hpp>
-#include <ngraph_ops/deconvolution_ie.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
+#include <legacy/ngraph_ops/deconvolution_ie.hpp>
 
 #include <ngraph/pattern/op/wrap_type.hpp>
 
@@ -20,7 +20,7 @@ NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertConvolutions, "ConvertConvolutions",
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertConvolution, "ConvertConvolution", 0);
 
 ngraph::pass::ConvertConvolution::ConvertConvolution() {
-    MATCHER_SCOPE(ConvertConvolution);
+    //MATCHER_SCOPE(ConvertConvolution);
     auto conv = ngraph::pattern::wrap_type<opset1::Convolution>();
 
     ngraph::matcher_pass_callback callback = [](pattern::Matcher& m) {
@@ -51,7 +51,7 @@ ngraph::pass::ConvertConvolution::ConvertConvolution() {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertGroupConvolution, "ConvertGroupConvolution", 0);
 
 ngraph::pass::ConvertGroupConvolution::ConvertGroupConvolution() {
-    MATCHER_SCOPE(ConvertGroupConvolution);
+    //MATCHER_SCOPE(ConvertGroupConvolution);
     auto gconv = ngraph::pattern::wrap_type<opset1::GroupConvolution>();
 
     ngraph::matcher_pass_callback callback = [](pattern::Matcher& m) {
@@ -98,7 +98,7 @@ ngraph::pass::ConvertGroupConvolution::ConvertGroupConvolution() {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertDeconvolution, "ConvertDeconvolution", 0);
 
 ngraph::pass::ConvertDeconvolution::ConvertDeconvolution() {
-    MATCHER_SCOPE(ConvertDeconvolution);
+    //MATCHER_SCOPE(ConvertDeconvolution);
     auto conv = ngraph::pattern::wrap_type<opset1::ConvolutionBackpropData>();
 
     ngraph::matcher_pass_callback callback = [](pattern::Matcher& m) {
@@ -132,7 +132,7 @@ ngraph::pass::ConvertDeconvolution::ConvertDeconvolution() {
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertGroupDeconvolution, "ConvertGroupDeconvolution", 0);
 
 ngraph::pass::ConvertGroupDeconvolution::ConvertGroupDeconvolution() {
-    MATCHER_SCOPE(ConvertGroupDeconvolution);
+    //MATCHER_SCOPE(ConvertGroupDeconvolution);
     auto gconv = ngraph::pattern::wrap_type<opset1::GroupConvolutionBackpropData>();
 
     ngraph::matcher_pass_callback callback = [](pattern::Matcher& m) {
