+++ inference-engine/tests/functional/inference_engine/transformations/convert_broadcast3_test.cpp
@@ -19,7 +19,7 @@
 #include <transformations/init_node_info.hpp>
 #include <ngraph/pass/visualize_tree.hpp>
 #include <transformations/op_conversions/convert_broadcast3.hpp>
-#include <ngraph_ops/convolution_ie.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
 #include <ngraph/pass/manager.hpp>
 
 #include "common_test_utils/ngraph_test_utils.hpp"
