+++ inference-engine/tests/functional/inference_engine/transformations/convert_convolution_test.cpp
@@ -18,8 +18,8 @@
 #include <transformations/utils/utils.hpp>
 #include <transformations/init_node_info.hpp>
 #include <ngraph/pass/visualize_tree.hpp>
-#include <transformations/op_conversions/convert_convolutions.hpp>
-#include <ngraph_ops/convolution_ie.hpp>
+#include <legacy/transformations/convert_convolutions.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
 #include <ngraph/pass/manager.hpp>
 
 #include "common_test_utils/ngraph_test_utils.hpp"
