+++ inference-engine/tests/functional/inference_engine/transformations/keep_constant_inputs_tests.cpp
@@ -9,7 +9,7 @@
 
 #include <ngraph/function.hpp>
 #include <ngraph/opsets/opset1.hpp>
-#include <ngraph_ops/convolution_ie.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
 #include <transformations/init_node_info.hpp>
 
 #include <ie_precision.hpp>
