+++ inference-engine/tests/functional/inference_engine/transformations/ngraph_1d_ops_reshape_test.cpp
@@ -13,7 +13,7 @@
 
 #include <ngraph/function.hpp>
 #include <ngraph/op/constant.hpp>
-#include <ngraph_ops/convolution_ie.hpp>
+#include <legacy/ngraph_ops/convolution_ie.hpp>
 #include <ngraph/pass/constant_folding.hpp>
 #include <legacy/transformations/convert_opset1_to_legacy/reshape_1d_ops.hpp>
 #include <transformations/init_node_info.hpp>
